# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusStatus = typeof HealthStatusStatus[keyof typeof HealthStatusStatus];


export const HealthStatusStatus = {
  ok: 'ok',
  degraded: 'degraded',
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
  database: boolean;
}

export interface LoginRequest {
  /** @minLength 1 */
  loginId: string;
  /** @minLength 8 */
  password: string;
}

export interface ChangePasswordRequest {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export interface CreateEmployeeRequest {
  /** Valid email address */
  email: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  lastName: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  employeeCode: string;
  /** @maxLength 30 */
  phone?: string;
  /** @maxLength 100 */
  department?: string;
  /** @maxLength 100 */
  jobTitle?: string;
  dateOfJoining: string;
}

export type UserResponseRole = typeof UserResponseRole[keyof typeof UserResponseRole];


export const UserResponseRole = {
  ADMIN: 'ADMIN',
  EMPLOYEE: 'EMPLOYEE',
} as const;

export interface UserResponse {
  id: string;
  loginId: string;
  /** Valid email address */
  email: string;
  role: UserResponseRole;
  mustChangePassword?: boolean;
}

export interface AuthResponse {
  token: string;
  user: UserResponse;
}

export type LeaveType = typeof LeaveType[keyof typeof LeaveType];


export const LeaveType = {
  ANNUAL: 'ANNUAL',
  SICK: 'SICK',
  CASUAL: 'CASUAL',
  UNPAID: 'UNPAID',
  OTHER: 'OTHER',
} as const;

export type LeaveStatus = typeof LeaveStatus[keyof typeof LeaveStatus];


export const LeaveStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
} as const;

export interface CreateLeaveRequest {
  leaveType: LeaveType;
  startDate: string;
  endDate: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  reason: string;
}

export type LeaveDecisionRequestStatus = typeof LeaveDecisionRequestStatus[keyof typeof LeaveDecisionRequestStatus];


export const LeaveDecisionRequestStatus = {
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
} as const;

export interface LeaveDecisionRequest {
  status: LeaveDecisionRequestStatus;
}

export interface LeaveEmployeeSummary {
  id: string;
  employeeCode: string;
  firstName: string;
  lastName: string;
  user: UserResponse;
}

export interface LeaveRequest {
  id: string;
  employeeId: string;
  leaveType: LeaveType;
  startDate: string;
  endDate: string;
  reason: string;
  status: LeaveStatus;
  reviewedAt?: string | null;
  reviewedById?: string | null;
  createdAt: string;
  updatedAt: string;
  employee?: LeaveEmployeeSummary;
}