# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication and account operations
  - name: employees
    description: Admin/HR employee management
  - name: leave
    description: Employee leave requests and Admin/HR decisions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "503":
          description: Server is running but the database is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with a Login ID or email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
        "400":
          description: Invalid password
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Employee accounts
    post:
      operationId: createEmployee
      tags: [employees]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          description: Employee account created
        "400":
          description: Invalid employee details
        "403":
          description: Admin/HR access required
  /leave-requests:
    post:
      operationId: createLeaveRequest
      tags: [leave]
      summary: Submit a leave request
      description: Employee-authenticated. Creates a pending leave request for the current user's employee profile.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveRequest"
      responses:
        "201":
          description: Leave request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
        "400":
          description: Invalid leave request details
        "401":
          description: Authentication required
        "403":
          description: Employee access required
        "404":
          description: Employee profile not found
    get:
      operationId: listLeaveRequests
      tags: [leave]
      summary: List all leave requests
      description: Admin/HR-authenticated. Each item includes the related employee summary.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: All leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
        "401":
          description: Authentication required
        "403":
          description: Admin/HR access required
  /leave-requests/me:
    get:
      operationId: listMyLeaveRequests
      tags: [leave]
      summary: List the current employee's leave requests
      description: Employee-authenticated. Returns leave requests for the current user's employee profile.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current employee's leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
        "401":
          description: Authentication required
        "403":
          description: Employee access required
        "404":
          description: Employee profile not found
  /leave-requests/{id}/status:
    patch:
      operationId: decideLeaveRequestStatus
      tags: [leave]
      summary: Approve or reject a pending leave request
      description: Admin/HR-authenticated. Only pending leave requests can be decided.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveDecisionRequest"
      responses:
        "200":
          description: Leave request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
        "400":
          description: Invalid decision
        "401":
          description: Authentication required
        "403":
          description: Admin/HR access required
        "404":
          description: Leave request not found
        "409":
          description: Leave request is not pending
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
          enum: [ok, degraded]
        database:
          type: boolean
      required:
        - status
        - database
    LoginRequest:
      type: object
      required: [loginId, password]
      properties:
        loginId: { type: string, minLength: 1 }
        password: { type: string, minLength: 8 }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 8 }
    CreateEmployeeRequest:
      type: object
      required: [email, firstName, lastName, employeeCode, dateOfJoining]
      properties:
        email: { type: string, description: Valid email address }
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
        employeeCode: { type: string, minLength: 1, maxLength: 40 }
        phone: { type: string, maxLength: 30 }
        department: { type: string, maxLength: 100 }
        jobTitle: { type: string, maxLength: 100 }
        dateOfJoining: { type: string, format: date }
    UserResponse:
      type: object
      required: [id, loginId, email, role]
      properties:
        id: { type: string }
        loginId: { type: string }
        email: { type: string, description: Valid email address }
        role: { type: string, enum: [ADMIN, EMPLOYEE] }
        mustChangePassword: { type: boolean }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/UserResponse" }
    LeaveType:
      type: string
      enum: [ANNUAL, SICK, CASUAL, UNPAID, OTHER]
    LeaveStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    CreateLeaveRequest:
      type: object
      required: [leaveType, startDate, endDate, reason]
      properties:
        leaveType: { $ref: "#/components/schemas/LeaveType" }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        reason: { type: string, minLength: 1, maxLength: 1000 }
    LeaveDecisionRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [APPROVED, REJECTED]
    LeaveEmployeeSummary:
      type: object
      required: [id, employeeCode, firstName, lastName, user]
      properties:
        id: { type: string }
        employeeCode: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        user: { $ref: "#/components/schemas/UserResponse" }
    LeaveRequest:
      type: object
      required:
        - id
        - employeeId
        - leaveType
        - startDate
        - endDate
        - reason
        - status
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        employeeId: { type: string }
        leaveType: { $ref: "#/components/schemas/LeaveType" }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        reason: { type: string }
        status: { $ref: "#/components/schemas/LeaveStatus" }
        reviewedAt: { type: string, format: date-time, nullable: true }
        reviewedById: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        employee: { $ref: "#/components/schemas/LeaveEmployeeSummary" }