# lib/api-zod/src/generated/types/createLeaveRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeaveType } from './leaveType';

export interface CreateLeaveRequest {
  leaveType: LeaveType;
  startDate: Date;
  endDate: Date;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  reason: string;
}