# lib/api-zod/src/generated/types/leaveDecisionRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeaveDecisionRequestStatus } from './leaveDecisionRequestStatus';

export interface LeaveDecisionRequest {
  status: LeaveDecisionRequestStatus;
}