# lib/api-zod/src/generated/types/leaveRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeaveEmployeeSummary } from './leaveEmployeeSummary';
import type { LeaveStatus } from './leaveStatus';
import type { LeaveType } from './leaveType';

export interface LeaveRequest {
  id: string;
  employeeId: string;
  leaveType: LeaveType;
  startDate: Date;
  endDate: Date;
  reason: string;
  status: LeaveStatus;
  reviewedAt?: Date | null;
  reviewedById?: string | null;
  createdAt: Date;
  updatedAt: Date;
  employee?: LeaveEmployeeSummary;
}