# lib/api-zod/src/generated/types/leaveType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type LeaveType = typeof LeaveType[keyof typeof LeaveType];


export const LeaveType = {
  ANNUAL: 'ANNUAL',
  SICK: 'SICK',
  CASUAL: 'CASUAL',
  UNPAID: 'UNPAID',
  OTHER: 'OTHER',
} as const;